Move to rdf4j instead of opensesame
opensesame has moved to eclipse.org under the new name rdf4j
