+++ server/src/main/java/com/thoughtworks/studios/shine/semweb/sesame/SesameURIReference.java
@@ -1,23 +1,23 @@
-/*************************GO-LICENSE-START*********************************
- * Copyright 2014 ThoughtWorks, Inc.
+/*
+ * Copyright 2018 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *    http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *************************GO-LICENSE-END***********************************/
+ */
 
 package com.thoughtworks.studios.shine.semweb.sesame;
 
 import com.thoughtworks.studios.shine.semweb.URIReference;
-import org.openrdf.model.URI;
+import org.eclipse.rdf4j.model.URI;
 
 public class SesameURIReference extends AbstractSesameResource implements URIReference {
     SesameURIReference(URI sesameNativeURI) {
